requirements.txt
fastapi
uvicorn
Pillow
transformers
torch
easyocr
numpy
webcolors
rembg
onnxruntime
